+++ packages/it-tests/package.json
@@ -31,6 +31,7 @@
 		"@remotion/google-fonts": "workspace:*",
 		"@remotion/serverless": "workspace:*",
 		"@remotion/studio-server": "workspace:*",
+		"@remotion/studio-shared": "workspace:*",
 		"@remotion/transitions": "workspace:*",
 		"@types/sharp": "^0.28.5",
 		"create-video": "workspace:*",
