+++ packages/it-tests/tsconfig.json
@@ -1,33 +1,35 @@
 {
-  "extends": "../tsconfig.settings.json",
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "noEmit": true,
-    "incremental": true
-  },
-  "references": [
-    {
-      "path": "../bundler"
-    },
-    {
-      "path": "../core"
-    },
-    {
-      "path": "../create-video"
-    },
-    {
-      "path": "../lambda"
-    },
-    {
-      "path": "../renderer"
-    },
-    {
-      "path": "../install-whisper-cpp"
-    },
-
-    {
-      "path": "../cli"
-    }
-  ],
-  "exclude": ["download-whisper.ts"]
+	"extends": "../tsconfig.settings.json",
+	"compilerOptions": {
+		"jsx": "react-jsx",
+		"noEmit": true,
+		"incremental": true
+	},
+	"references": [
+		{
+			"path": "../bundler"
+		},
+		{
+			"path": "../core"
+		},
+		{
+			"path": "../create-video"
+		},
+		{
+			"path": "../lambda"
+		},
+		{
+			"path": "../renderer"
+		},
+		{
+			"path": "../install-whisper-cpp"
+		},
+		{
+			"path": "../studio-shared"
+		},
+		{
+			"path": "../cli"
+		}
+	],
+	"exclude": ["download-whisper.ts"]
 }
